BACKEND/requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
email-validator
python-dotenv